rename api library from 'pkg' to 'sdk' to avoid conflicts with release binaries
