+++ internal/pkg/cache/pack.go
@@ -5,7 +5,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/hashicorp/nomad-pack/pkg/pack"
+	"github.com/hashicorp/nomad-pack/sdk/pack"
 )
 
 // PackConfig represents the common configuration required by all packs. Used primarily
