+++ internal/pkg/loader/loader.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/hcl/v2/hclsimple"
-	"github.com/hashicorp/nomad-pack/pkg/pack"
+	"github.com/hashicorp/nomad-pack/sdk/pack"
 )
 
 func Load(name string) (*pack.Pack, error) {
