+++ sdk/pack/metadata_test.go
